+++ modules/top-level.nix
@@ -46,18 +46,48 @@ let types = lib.types; in
     buildCommand = lib.mkOption {
       default = null;
       type = types.nullOr types.str;
-      description = "Command to build the jar application. If not provided, a default builder is used";
+      description = lib.mdDoc "Command to build the jar application. If not provided, a default builder is used";
     };
+
+    builder-extra-inputs = lib.mkOption {
+      default = [ ];
+      type = types.listOf types.package;
+      description = lib.mdDoc "Extra inputs to the default builder";
+    };
+
+    builder-java-opts = lib.mkOption {
+      type = types.listOf types.str;
+      default = [ ];
+      description = lib.mdDoc "List of Java options to include in default builder command";
+    };
+
+    builder-preBuild = lib.mkOption {
+      default = "";
+      type = types.str;
+      description = lib.mdDoc "Pre build commands for the default builder";
+    };
+    builder-postBuild = lib.mkOption {
+      default = "";
+      type = types.str;
+      description = lib.mdDoc "Post build commands for the default builder";
+    };
+
 
     lockfile = lib.mkOption {
       default = null;
-      type = types.nullOr types.str;
+      type = types.nullOr types.path;
       description = lib.mdDoc "The lock file";
     };
 
     compileCljOpts = lib.mkOption {
       default = null;
       description = lib.mdDoc "Override options passed to compile-clj";
+      type = types.anything;
+    };
+
+    aliases = lib.mkOption {
+      default = null;
+      description = lib.mdDoc "Aliases used when calculating basis and src-dirs";
       type = types.anything;
     };