+++ src/cljnix/builder_cli.clj
@@ -5,18 +5,19 @@
     [cljnix.check :as check]
     [clojure.data.json :as json]))
 
-(defn- check-main-class
-  [args]
-  (or
-   (check/main-gen-class
-    (interleave
-     [:lib-name :version :main-ns]
-     args))
-   (throw (ex-info "main-ns class does not specify :gen-class" {:args args}))))
-
 (defn- str->json
   [s]
   (json/read-str s :key-fn keyword))
+
+(defn- check-main-class
+  [args]
+  (or
+   (-> (zipmap [:lib-name :version :main-ns :compile-clj-opts :javac-opts :aliases] args)
+          (update :compile-clj-opts str->json)
+          (update :javac-opts str->json)
+          (update :aliases str->json)
+          (check/main-gen-class))
+   (throw (ex-info "main-ns class does not specify :gen-class" {:args args}))))
 
 ; Internal CLI helpers
 (defn -main
@@ -26,17 +27,19 @@
     (apply utils/expand-shas! args)
 
     (= cmd "jar")
-    (build/jar
-      (interleave
-        [:lib-name :version]
-        args))
+    (-> (zipmap [:lib-name :version :main-ns :compile-clj-opts :javac-opts :aliases] args)
+        (update :compile-clj-opts str->json)
+        (update :javac-opts str->json)
+        (update :aliases str->json)
+        (build/jar))
 
     (= cmd "uber")
     (do
       (check-main-class args)
-      (-> (zipmap [:lib-name :version :main-ns :compile-clj-opts :javac-opts] args)
+      (-> (zipmap [:lib-name :version :main-ns :compile-clj-opts :javac-opts :aliases] args)
           (update :compile-clj-opts str->json)
           (update :javac-opts str->json)
+          (update :aliases str->json)
           (build/uber)))
 
     (= cmd "check-main")