+++ src/cljnix/utils.clj
@@ -10,19 +10,20 @@
     [clojure.tools.gitlibs.impl :as gli]
     [clojure.tools.deps :as deps]
     [babashka.fs :as fs]
+    [babashka.http-client :as http]
     [version-clj.core :as version]
     [clojure.zip :as zip]
     [borkdude.rewrite-edn :as r]
-    [clojure.tools.deps.util.io :refer [printerrln]])
+    [clojure.tools.deps.util.io :refer [printerrln]]
+    [clojure.tools.deps.extensions.pom :refer [read-model-file]])
   (:import
     [org.apache.maven.model.io.xpp3 MavenXpp3Reader]
-    [org.apache.maven.model Model]))
+    [org.apache.maven.model Model Repository Dependency]
+    [java.net URI]))
 
 (defn throw+
   [msg data]
   (throw (ex-info (prn-str msg data) data)))
-
-(def ^:dynamic *mvn-repos* mvn/standard-repos)
 
 (defn- snapshot?
   [path]
@@ -34,12 +35,21 @@
       (string/includes? "snapshot")))
 
 
-; https://maven.apache.org/ref/3.6.3/maven-model/apidocs/index.html
+; https://maven.apache.org/ref/3.9.6/maven-model/apidocs/index.html
 (defn- pom
   ^Model [pom-path]
+  (try
+    (when (= "pom" (fs/extension pom-path))
+      (let [f (io/input-stream (str pom-path))]
+        (.read (MavenXpp3Reader.) f)))
+    (catch Exception _
+      (printerrln "Error parsing [skiping]" pom-path))))
+
+; Alternative implementation
+(defn- pom'
+  ^Model [pom-path]
   (when (= "pom" (fs/extension pom-path))
-    (let [f (io/input-stream (str pom-path))]
-      (.read (MavenXpp3Reader.) f))))
+    (read-model-file (io/file (str pom-path)) {:mvn/repos mvn/standard-repos})))
 
 ;; Snapshot jar can be
 ;; foo-123122312.jar
@@ -83,53 +93,66 @@
                                               ext))))
        :snapshot (str artifact-id "-" snapshot-version "." ext)})))
 
-
-(defn get-parent
- [pom-path]
- (when-let [parent (some-> (pom pom-path) (.getParent))]
-    (let [parent-path
-          (fs/path
-            @mvn/cached-local-repo
-            (string/replace (.getGroupId parent) "." "/")
-            (.getArtifactId parent)
-            (.getVersion parent)
-            (format "%s-%s.pom" (.getArtifactId parent) (.getVersion parent)))]
-      (when (fs/exists? parent-path)
-        (str parent-path)))))
-
-
 (defn- get-mvn-repo-name
-  [path]
-  (let [info-file (fs/path (fs/parent path) "_remote.repositories")
-        file-name (fs/file-name path)
-        repo-name-finder (fn [s] (second (re-find
-                                           (re-pattern (str file-name #">(\S+)="))
-                                           s)))]
-    (some #(let [repo-name (repo-name-finder %)]
-             (when (contains? *mvn-repos* repo-name)
-               repo-name))
-          (fs/read-all-lines info-file))))
+  [path mvn-repos]
+  (try
+    (let [info-file (fs/path (fs/parent path) "_remote.repositories")
+          file-name (fs/file-name path)
+          repo-name-finder (fn [s] (second (re-find
+                                             (re-pattern (str file-name #">(\S+)="))
+                                             s)))]
+      (some #(let [repo-name (repo-name-finder %)]
+               (when (contains? mvn-repos repo-name)
+                 repo-name))
+           (fs/read-all-lines info-file)))
+    (catch Exception _ nil)))
+
+
+(defn- valid-url?
+  [url]
+  (try
+    (boolean (http/head url))
+    (catch Exception _ false)))
+
+(defn- join-url
+  [url path]
+  (let [url (if (string/ends-with? url "/")
+              url
+              (str url "/"))]
+    (-> (URI. url)
+      (.resolve path)
+      (str))))
+
+(defn- find-mvn-repo-name!
+  [mvn-path mvn-repos]
+  (some (fn [[repo-name {:keys [url]}]]
+          (when (valid-url? (join-url url mvn-path))
+            repo-name))
+        mvn-repos))
+
 
 (defn mvn-repo-info
   "Given a path for a jar in the maven local repo, e.g:
    $REPO/babashka/fs/0.1.4/fs-0.1.4.jar
    return the maven repository url and the dependecy url"
-  [path {:keys [cache-dir exact-version]
-         :or {cache-dir @mvn/cached-local-repo}}]
-  {:pre [(if (snapshot? path)
-           (not (nil? exact-version))
-           true)]}
+  [path & {:keys [cache-dir exact-version mvn-repos]
+           :or {cache-dir @mvn/cached-local-repo
+                mvn-repos mvn/standard-repos}}]
   (let [{:keys [resolved-path snapshot]} (resolve-snapshot path exact-version)
-        repo-name (get-mvn-repo-name resolved-path)
-        repo-url (get-in *mvn-repos* [repo-name :url])
+        mvn-path (str (fs/relativize cache-dir resolved-path))
+        repo-name (or
+                    (get-mvn-repo-name resolved-path mvn-repos)
+                    (find-mvn-repo-name! mvn-path mvn-repos))
+        repo-url (get-in mvn-repos [repo-name :url])
         repo-url (cond
                    (nil? repo-url) (throw+ "Maven repo not found"
-                                           {:mvn-repos *mvn-repos*
+                                           {:mvn-repos mvn-repos
+                                            :repo-name repo-name
                                             :file path})
                    ((complement string/ends-with?) repo-url "/") (str repo-url "/")
                    :else repo-url)]
      (cond-> {:mvn-repo repo-url
-              :mvn-path (str (fs/relativize cache-dir resolved-path))
+              :mvn-path mvn-path
               :url (str repo-url (fs/relativize cache-dir resolved-path))}
        snapshot (assoc :snapshot snapshot))))
 
@@ -250,23 +273,134 @@
     (string/split #"/")
     (->> (apply keyword))))
 
-(comment
-  (expand-shas! "/home/jlle/projects/clojure-lsp")
 
-  (mvn-repo-info "/home/jlle/.m2/repository/org/clojure/clojure/1.11.1/clojure-1.11.1.jar"
-                 @mvn/cached-local-repo)
-  (mvn-repo-info "/home/jlle/.m2/repository/org/clojure/pom.contrib/1.1.0/pom.contrib-1.1.0.pom"
-                 @mvn/cached-local-repo)
-  (mvn-repo-info "/tmp/clj-cache7291733112505590287/mvn/org/clojure/pom.contrib/1.1.0/pom.contrib-1.1.0.pom"
-                 @mvn/cached-local-repo)
+(defn mvn?
+  [[_ {:keys [mvn/version]}]]
+  (boolean version))
+
+(defn git?
+  [[_ {:keys [git/url]}]]
+  (boolean url))
+
+(defn artifact->pom
+  [path]
+  (str (->> (fs/glob (fs/parent path) "*.pom")
+            (sort (fn [x y]
+                    (cond
+                      (string/includes? (fs/file-name x) "SNAPSHOT") -1
+                      (string/includes? (fs/file-name y) "SNAPSHOT") 1
+                      :else (compare (fs/file-name x) (fs/file-name y)))))
+            first)))
+
+(defn get-repos
+  "For a given path to a pom, returns all maven repos"
+  [pom-path]
+  (try
+    (let [pom (io/file pom-path)
+          model (read-model-file pom nil)]
+      (into {}
+            (map (fn [^Repository repo] (vector (.getId repo)
+                                                {:url (.getUrl repo)})))
+            (.getRepositories model)))
+    (catch Exception _ {})))
+
+
+(defn get-maven-repos
+  "Given a basis, returns all maven repos"
+  [basis]
+  (into (merge mvn/standard-repos (:mvn/repos basis))
+        (comp
+          (filter mvn?)
+          (map val)
+          (mapcat :paths)
+          (map artifact->pom)
+          (map get-repos))
+        (:libs basis)))
+
+
+(defn- dep-path
+  [^Dependency dep]
+  (str
+    (fs/path
+      @mvn/cached-local-repo
+      (string/replace (.getGroupId dep) "." "/")
+      (.getArtifactId dep)
+      (.getVersion dep)
+      (format "%s-%s.pom" (.getArtifactId dep) (.getVersion dep)))))
+
+
+(defn- get-deps
+  [pom-path]
+  (some->> (pom pom-path)
+           (.getDependencyManagement)
+           (.getDependencies)
+           (map dep-path)
+           (filter fs/exists?)
+           (distinct)
+           (remove #(= % pom-path))))
+
+(defn get-management-deps
+  "Given a pom file, recursively returns all the management dependencies
+   In many cases those POMs contain BOM information"
+  [pom-path]
+  (when pom-path
+    (loop [deps #{}
+           deps' (get-deps pom-path)]
+      (let [new-deps (remove deps deps')]
+        (if-let [dep (first new-deps)]
+          (recur (conj deps dep)
+                 (concat (rest new-deps) (get-deps dep)))
+          deps)))))
+
+
+(defn get-parent
+ [pom-path]
+ (when-let [parent (some-> (pom pom-path) (.getParent))]
+    (let [parent-path
+          (fs/path
+            @mvn/cached-local-repo
+            (string/replace (.getGroupId parent) "." "/")
+            (.getArtifactId parent)
+            (.getVersion parent)
+            (format "%s-%s.pom" (.getArtifactId parent) (.getVersion parent)))]
+      (when (fs/exists? parent-path)
+        (str parent-path)))))
+
+
+(defn get-parent-poms
+  "Given a pom file, recursively returns all parent POMs"
+  [pom-path]
+  (loop [parents #{}
+         new-parent (get-parent pom-path)]
+     (if (nil? new-parent)
+       parents
+       (recur (conj parents new-parent)
+              (get-parent new-parent)))))
+
+(comment
+
+  (get-maven-repos
+    (deps/create-basis {:user nil
+                        :project (str (fs/expand-home "~/projects/clj-demo-project/deps.edn"))}))
+
+  (expand-shas! (fs/expand-home "~/projects/clojure-lsp"))
+
+  (mvn-repo-info (fs/expand-home "~/.m2/repository/org/clojure/clojure/1.11.1/clojure-1.11.1.jar"))
+  (mvn-repo-info (fs/expand-home "~/.m2/repository/org/clojure/pom.contrib/1.1.0/pom.contrib-1.1.0.pom"))
+  (get-management-deps (fs/expand-home "~/.m2/repository/metosin/reitit/0.5.15/reitit-0.5.15.pom"))
+
+  (fs/expand-home "~/.m2/repository/metosin/reitit/0.5.15/reitit-0.5.15.pom")
+  (get-management-deps (fs/expand-home "~/.m2/repository/com/google/firebase/firebase-admin/9.2.0/firebase-admin-9.2.0.pom"))
+  (get-parent-poms (fs/expand-home "~/.m2/repository/io/grpc/grpc-bom/1.55.1/pom.xml"))
 
 
   (=
     (mvn-repo-info
-      "/home/jlle/.m2/repository/clj-kondo/clj-kondo/2022.04.26-SNAPSHOT/clj-kondo-2022.04.26-SNAPSHOT.jar"
+      (fs/expand-home "~/.m2/repository/clj-kondo/clj-kondo/2022.04.26-SNAPSHOT/clj-kondo-2022.04.26-SNAPSHOT.jar")
       {:exact-version "2022.04.26-SNAPSHOT"})
+
     (mvn-repo-info
-      "/home/jlle/.m2/repository/clj-kondo/clj-kondo/2022.04.26-SNAPSHOT/clj-kondo-2022.04.26-20220502.201054-5.jar"
+      (fs/expand-home "~/.m2/repository/clj-kondo/clj-kondo/2022.04.26-SNAPSHOT/clj-kondo-2022.04.26-20220502.201054-5.jar")
       {:exact-version "2022.04.26-20220502.201054-5"}))